* Smiley editing * File security fix

+++ inc/model/files/image.php
@@ -275,7 +275,7 @@ class image extends \fpcm\model\abstracts\file {
      */
     public function save()
     {
-        if ($this->exists(true)) {
+        if ($this->exists(true) || !$this->isValidDataFolder(dirname($this->fullpath))) {
             return false;
         }
 
@@ -291,7 +291,7 @@ class image extends \fpcm\model\abstracts\file {
      */
     public function update()
     {
-        if (!$this->exists(true)) {
+        if (!$this->exists(true) || !$this->isValidDataFolder(dirname($this->fullpath))) {
             return false;
         }
 
@@ -310,6 +310,10 @@ class image extends \fpcm\model\abstracts\file {
      */
     public function delete()
     {
+        if (!$this->isValidDataFolder(dirname($this->fullpath))) {
+            return false;
+        }
+        
         parent::delete();
         if (file_exists($this->getFileManagerThumbnail())) {
             unlink($this->getFileManagerThumbnail());
@@ -379,6 +383,10 @@ class image extends \fpcm\model\abstracts\file {
             return $count > 0 ? true : false;
         }
 
+        if (!$this->isValidDataFolder(dirname($this->fullpath))) {
+            return false;
+        }
+
         return (parent::exists() && $count > 0) ? true : false;
     }
 
@@ -389,6 +397,10 @@ class image extends \fpcm\model\abstracts\file {
      */
     public function existsFolder()
     {
+        if (!$this->isValidDataFolder(dirname($this->fullpath))) {
+            return false;
+        }
+
         return parent::exists();
     }
